+++ src/test/java/wordle/view/InputView.java
@@ -1,0 +1,5 @@
+package wordle.view;
+
+public interface InputView {
+    String inputWord();
+}
